Move logging into inner method for reorder pass
The reason for inner/outer method is to keep the outer method conforming
to the typedef for a comms graph pass which returns one obj, while
allowing unit tests to call the inner method that returns more metadata
useful for testing the pass.  The logs should be in the inner part, so
they are functional also during unit testing.

ghstack-source-id: ef8170a91bc4a8bb49a668ca77c79a448f379b3a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156879