ci: Update lint workflow to only run on changed files for PRs
This modifies the lint workflow to use the new get-changed-files
workflow to optimize lint execution by only running on files
that have actually changed in pull requests.

This more closely mirrors the type of behavior that users
expect when running lint locally on their PRs.

This also leaves the default behavior as a fallback for when
you're not running on a pull request.

Since lint runs on the pull_request event I'm not really worried about
any type of ciflow shenanigans in this.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 2f0674a53fe08e7a765062c84578088cccae8f28
Pull-Request: https://github.com/pytorch/pytorch/pull/158518
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>