[cpp_wrapper] Build main and kernel code in separate threads
The main change of this commit is building cpp_wrapper code asynchronously, with a thread for the wrapper code, a thread for the kernel code, and a final thread to link. This improves build speeds, even while enabling an LTO link to improve performance.

Additionally, significantly reduce the cold-start time for detecting compatible precompiled headers. We had accidentally preprocessed the underlying header in -O3, which is deeply unnecessary to simply detect content changes.

ghstack-source-id: ad74c566c2ed8dc9ea3d63bc0882c3c3b70673aa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/154551