[C10D] fix slow init due to repeated dns resolution failure
It can be be very slow to repeatedly hit DNS resolution failure, but
its very helpful to have DNS names in logs by default. So we try to use DNS
but if we hit a transient failure we just disable it for the remainder of the
job, logging IP addresses instead.

Fixes #159007

ghstack-source-id: b3574b5c9dfd8373ef3833c9edb3e142696041a6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159596