[AOTI][experiment] Provide an option to cross-compile for Windows
Summary: Because PT2 doesn't run on native Windows due to the missing Triton support, we take a cross-compilation path for AOTInductor. Concretely, we can run AOTI on WSL, specifying package_cpp_only to generate a C++ project with cpp files for the main wrapper and weights and generate ptx files for kernels. Then we take the C++ project files to native Windows and compile it with native Windows toolchain. This PR adds a cross-compilation config so that WSL export can work. Next PR in this stack will update AOTI runner code so that they can work with this flow.
ghstack-source-id: 4c4ed582957eaa640ce67794feedca2258d46e24
Pull-Request: https://github.com/pytorch/pytorch/pull/159794