[pt2e] Avoid getting model device once per node
**Summary:** Previously, we call `assert_and_get_unqiue_device`
once per node in both prepare and convert. This is expensive and
unnecessary since the model device is the same across all nodes,
so we should just call this once in the beginning and reuse the
same model device across all the nodes.

**Test Plan:**
python test/test_quantization.py -k TestQuantizePT2E