[DTensor] Support user-supplied Generator for random ops
If the user provides a generator kwarg to a random op (e.g.
nn.init.uniform_(..., generator=my_generator)), we can still advance
that generator's state in a SPMD-global way so that each local-tensor
gets appropriate values and the generator advances to the same state as
if it had operated on the full tensor.

ghstack-source-id: d2523f2659c4ff15aab35fdf322b332dca46694b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159933