[dynamo] Refactor transform() so that instruction translator can be used as a tracing function. [2/n]
Pull Request resolved: https://github.com/pytorch/pytorch/pull/160815

We are refactoring dynamo code for convert frame so that we can have modularized pieces sharable between different compiler frontends (e.g. torch.compile, precompile and torch.export).

This PR follows the last one which separate out the part to run instruction translator on a given frame and return a DynamoTracerOutput.

The end result is a free function that runs instruction translator indepedently. A follow up diff will wrap the low level function.
ghstack-source-id: 303652896
@exported-using-ghexport

Differential Revision: [D80388694](https://our.internmc.facebook.com/intern/diff/D80388694/)