[dynamo] Refactor convert_frame.compile_frame to be self contained function. [5/n]
Pull Request resolved: https://github.com/pytorch/pytorch/pull/160900

convert_frame.compile_frame used to take a callback transform function which will capture the frame object it has, but the frame information is not passed directly into compile_frame function.

This PR changes the signature of compile_frame so that frame information is directly passed in the function without taking a callback. This makes it easier to build fullgraph capture API on top of compile_frame.
ghstack-source-id: 304836212
@exported-using-ghexport

Differential Revision: [D80469801](https://our.internmc.facebook.com/intern/diff/D80469801/)